% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{add_hierarchy_column}
\alias{add_hierarchy_column}
\title{Add Taxonomic Hierarchy Column to a Dataset}
\usage{
add_hierarchy_column(data, data_dwca = NULL, verbose = TRUE)
}
\arguments{
\item{data}{A data frame containing taxonomic information. Must include columns \code{taxonId},
\code{parentNameUsageID}, and \code{scientificName}.}

\item{data_dwca}{A data frame representing the complete taxonomic dataset (e.g., full taxonomy table).
Must include the same columns as \code{data}. Defaults to \code{data} if not provided.}

\item{verbose}{Logical. If TRUE, the function will print additional messages to provide feedback on its progress. Default is TRUE.}
}
\value{
A data frame with an additional \code{hierarchy} column. The \code{hierarchy} column contains
a string representing the lineage of parent \code{scientificName} values, separated by " > ".
}
\description{
This function adds a new column, \code{hierarchy}, to a dataset, where each row contains
the taxonomic hierarchy of parent \code{scientificName} values for the given \code{taxonId}.
The hierarchy is constructed using a recursive helper function.
}
\examples{
# Example dataset
data <- dplyr::tibble(
  taxonId = c("1", "2", "3"),
  parentNameUsageID = c(NA, "1", "2"),
  scientificName = c("Kingdom", "Phylum", "Class")
)

# Add hierarchy column
data_with_hierarchy <- SHARK4R:::add_hierarchy_column(data)
print(data_with_hierarchy)

}
\keyword{internal}
