% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{add_species_info}
\alias{add_species_info}
\title{Helper function to add species information to species list}
\usage{
add_species_info(shark_species_list, species)
}
\arguments{
\item{shark_species_list}{A data frame containing shark species information.}

\item{species}{A data frame containing gathered species information.}
}
\value{
A data frame containing the shark species list with added species information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This internal function adds species information to the shark species list.
}
\keyword{internal}
