% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms_api_functions.R
\name{add_worms_taxonomy}
\alias{add_worms_taxonomy}
\title{Retrieve and Organize WoRMS Taxonomy for SHARK Aphia IDs}
\usage{
add_worms_taxonomy(aphia_id, scientific_name = NULL, verbose = TRUE)
}
\arguments{
\item{aphia_id}{A numeric vector containing Aphia IDs for which WoRMS taxonomy needs to be updated.}

\item{scientific_name}{A character vector of scientific names. If provided, Aphia IDs will be retrieved from the scientific names. The length of \code{scientific_name} must match the length of \code{aphia_id}. Defaults to \code{NULL}, in which case the function will only add taxonomy to the provided Aphia IDs.}

\item{verbose}{A logical indicating whether to print progress messages. Default is TRUE.}
}
\value{
A data frame containing updated WoRMS taxonomy information.
}
\description{
This function collects WoRMS (World Register of Marine Species) taxonomy information for a given set of Aphia IDs.
The data is organized into a full taxonomic table that can be joined with data downloaded from \href{https://shark.smhi.se/}{SHARK}.
}
\examples{
\dontrun{
# Update WoRMS taxonomy for a set of Aphia IDs
updated_taxonomy <- add_worms_taxonomy(c(149619, 149122, 11))
print(updated_taxonomy)

# Update WoRMS with an unknown Aphia ID and scientific names
with_names <- add_worms_taxonomy(c(149619, 149122, 11, NA),
                                 c("Cerataulina pelagica",
                                 "Chaetoceros didymus",
                                 "Ciliophora",
                                 "Dinophysis"))
print(with_names)
}

}
\seealso{
https://cran.r-project.org/web/packages/worrms/index.html

\code{\link{download_sharkdata}}, \code{\link{update_dyntaxa_taxonomy}}, \href{http://www.marinespecies.org/rest/}{WoRMS API Documentation}
}
