% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms_api_functions.R
\name{assign_plankton_groups}
\alias{assign_plankton_groups}
\title{Assign Plankton Group to Scientific Names}
\usage{
assign_plankton_groups(
  scientific_names,
  aphia_ids = NULL,
  diatom_class = "Bacillariophyceae",
  dinoflagellate_class = "Dinophyceae",
  cyanobacteria_class = "Cyanophyceae",
  cyanobacteria_phylum = "Cyanobacteria",
  match_first_word = TRUE,
  marine_only = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{scientific_names}{A character vector of scientific names of marine species.}

\item{aphia_ids}{A numeric vector of Aphia IDs corresponding to the scientific names. If provided, it improves the accuracy of the matching process. The length of \code{aphia_ids} must match the length of \code{scientific_names}. Defaults to \code{NULL}, in which case the function will attempt to assign plankton groups based only on the scientific names.}

\item{diatom_class}{A character string representing the diatom class. Default is "Bacillariophyceae".}

\item{dinoflagellate_class}{A character string representing the dinoflagellate class. Default is "Dinophyceae".}

\item{cyanobacteria_class}{A character string representing the cyanobacteria class. Default is "Cyanophyceae".}

\item{cyanobacteria_phylum}{A character string representing the cyanobacteria phylum. Default is "Cyanobacteria".}

\item{match_first_word}{A logical value indicating whether to match the first word of the scientific name if the Aphia ID is missing. Default is TRUE.}

\item{marine_only}{A logical value indicating whether to restrict the results to marine taxa only. Default is \code{FALSE}.}

\item{verbose}{A logical value indicating whether to print progress messages. Default is TRUE.}
}
\value{
A data frame with two columns: \code{scientific_name} and \code{plankton_group}, where the plankton group is assigned based on taxonomic classification.
}
\description{
This function assigns plankton groups (e.g., Diatoms, Dinoflagellates, Cyanobacteria, or Other)
to a list of scientific names or Aphia IDs by retrieving species information from the
World Register of Marine Species (WoRMS). The function checks both Aphia IDs and scientific names,
handles missing records, and assigns the appropriate plankton group based on taxonomic classification in WoRMS.
}
\details{
The \code{aphia_ids} parameter is not necessary but, if provided, will improve the certainty of the
matching process. If \code{aphia_ids} are available, they will be used directly to retrieve more accurate
WoRMS records. If missing, the function will attempt to match the scientific names to Aphia IDs by
querying WoRMS using the scientific name(s), with an additional fallback mechanism to match based on the
first word of the scientific name.
}
\examples{
# Assign plankton groups to a list of species
result <- assign_plankton_groups(
  scientific_names = c("Tripos fusus", "Diatoma", "Nodularia spumigena", "Octactis speculum"),
  aphia_ids = c(840626, 149013, 160566, NA)
)

}
\seealso{
\url{https://cran.r-project.org/web/packages/worrms/index.html}
}
