% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_codes.R
\name{check_code_proj}
\alias{check_code_proj}
\title{Check matches of reported project (PROJ) codes in SMHIs codelist}
\usage{
check_code_proj(data)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
unmatched codes with true or false results
}
\description{
Check matches of reported project (PROJ) codes in SMHIs codelist
}
