% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{check_data_version}
\alias{check_data_version}
\title{Check if dataset versions are up to date}
\usage{
check_data_version(dataset_file_name)
}
\arguments{
\item{dataset_file_name}{A character vector containing the dataset file names (with the .zip extension).}
}
\value{
A data frame indicating the status of dataset versions, including whether they are up to date.
}
\description{
This function checks the status of dataset versions by comparing them with the latest available versions.
}
\details{
This function compares the provided dataset file names with the latest versions available in the \href{https://sharkdata.smhi.se/}{SHARKdata} repository.
It returns a data frame with information about each dataset, specifying whether it is up to date or not.
}
\examples{
# Example with a list of dataset file names
\dontrun{
datasets <- load_dataset_names("Phytoplankton")
dataset_names <- unique(datasets$dataset_name[1:5])
data <- download_sharkdata(dataset_names)
status_list <- check_data_version(data$dataset_file_name)
}

}
\seealso{
\code{\link{load_dataset_names}}, \code{\link{download_sharkdata}}, \code{\link{update_data}}, \href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
