% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fields.R
\name{check_HarbourPorpoise}
\alias{check_HarbourPorpoise}
\title{Check if the required and recommended datatype-specific SHARK system fields are present}
\usage{
check_HarbourPorpoise(data, level = "error")
}
\arguments{
\item{data}{The data frame.}

\item{level}{The level of error reporting, i.e. "error" or "warning". Recommended fields are only checked in case of "warning".}
}
\value{
Any warnings or errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated and has been replaced by \code{\link[=check_fields]{check_fields()}}.
}
\keyword{internal}
