% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_phytoplankton_carbon}
\alias{check_phytoplankton_carbon}
\title{Check range of data for specific parameters
Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype}
\usage{
check_phytoplankton_carbon(data)

check_phytoplankton_carbon(data)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
ggplot and tibble of data with outliers

ggplot and tibble of data with outliers
}
\description{
Check range of data for specific parameters
Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype

Check range of data for specific parameters
Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype
}
