% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logical.R
\name{check_value_logical}
\alias{check_value_logical}
\title{Check logical assumptions of data for specific variables and parameters}
\usage{
check_value_logical(data)
}
\arguments{
\item{data}{A data frame. Must contain a column named \code{value}.}
}
\value{
A DT datatable with problematic entries, or NULL (invisibly) if all values are valid.
}
\description{
Check logical assumptions of data for specific variables and parameters
}
