% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logical.R
\name{check_zero_value}
\alias{check_zero_value}
\title{Check logical assumptions of data for specific variables and parameters}
\usage{
check_zero_value(data)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
logical test with true or false results
}
\description{
Check logical assumptions of data for specific variables and parameters
}
