% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_zoobenthos_counted}
\alias{check_zoobenthos_counted}
\title{Uses data from national marine monitoring for the last 5 years to identify outliers}
\usage{
check_zoobenthos_counted(data)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
tibble of data with outliers
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated and has been replaced by \code{\link[=check_outliers]{check_outliers()}}.

Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype
}
\keyword{internal}
