% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_zooplankton_abund}
\alias{check_zooplankton_abund}
\title{Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype}
\usage{
check_zooplankton_abund(data)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
tibble of data with outliers
}
\description{
Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype
}
