% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{construct_dyntaxa_missing_table}
\alias{construct_dyntaxa_missing_table}
\title{Construct Dyntaxa taxonomy table from individually API calls}
\usage{
construct_dyntaxa_missing_table(
  parent_ids,
  subscription_key = Sys.getenv("DYNTAXA_KEY"),
  shark_output = TRUE,
  recommended_only = TRUE,
  add_genus_children = FALSE,
  drop_morphotypes = TRUE,
  add_hierarchy = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{parent_ids}{A list or vector of parent taxon IDs for which taxonomy information is requested. These IDs must be valid
according to the Dyntaxa API and can be a combination of single or multiple IDs.}

\item{subscription_key}{A Dyntaxa API subscription key. By default, the key
is read from the environment variable \code{DYNTAXA_KEY}.

You can provide the key in three ways:
\itemize{
\item \strong{Directly as a parameter}:
\code{construct_dyntaxa_missing_table(list(c(5000055, 6011755)), subscription_key = "your_key_here")}
\item \strong{Temporarily for the session}:
\code{Sys.setenv(DYNTAXA_KEY = "your_key_here")}
\item \strong{Permanently across sessions} by adding it to your \code{~/.Renviron} file.
Use \code{usethis::edit_r_environ()} to open the file, then add:
\code{DYNTAXA_KEY=your_key_here}
}}

\item{shark_output}{Logical. If TRUE, the function will return columns formatted to match the SHARK data submission format.
If FALSE, it will return a broader set of taxonomy information. Defaults to TRUE.}

\item{recommended_only}{Logical. If TRUE, only recommended (accepted) names will be returned. Defaults to TRUE.}

\item{add_genus_children}{Logical. If TRUE, the function will include all valid genus-level children for each parent taxon.
Defaults to FALSE.}

\item{drop_morphotypes}{Logical. If TRUE, taxa with the rank of "MorphoType" will be excluded from the output to simplify
the taxonomy structure. Defaults to TRUE.}

\item{add_hierarchy}{Logical. If TRUE, the hierarchical relationship of taxa (from parent to child) will be included in the output.
Defaults to FALSE.}

\item{verbose}{Logical. If TRUE, a progress bar will be displayed during execution. Defaults to TRUE.}
}
\value{
A data frame with taxonomy information, including taxon IDs, parent IDs, ranks, names, and other details.
The data frame may also include hierarchical information depending on the parameters set.
}
\description{
This internal function constructs a taxonomy table by individually querying the SLU Artdatabanken API (Dyntaxa)
using a list of parent taxon IDs. It fetches taxonomy information for the provided taxon IDs and
organizes them into a hierarchical structure. The function is capable of filtering based on
recommended (accepted) names, handling genus-related children, and formatting the results for SHARK output.
}
\details{
A valid Dyntaxa API subscription key is required.
You can request a free key for the "Taxonomy" service from the ArtDatabanken API portal:
\url{https://api-portal.artdatabanken.se/}

\strong{Note}: Please review the \href{https://www.slu.se/artdatabanken/rapportering-och-fynd/oppna-data-och-apier/}{API conditions}
and \href{https://api-portal.artdatabanken.se/}{register for access} before using the API. Data collected through the API
is stored at SLU Artdatabanken. Please also note that the authors of \code{SHARK4R} are not affiliated with SLU Artdatabanken.
}
\examples{
\dontrun{
# Construct Dyntaxa taxonomy table for taxon IDs 238366 and 1010380
parent_ids <- get_dyntaxa_parent_ids(c(238366, 1010380), "your_subscription_key")
taxonomy_table <- SHARK4R:::construct_dyntaxa_missing_table(parent_ids, "your_subscription_key")
print(taxonomy_table)
}

}
\keyword{internal}
