% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{construct_dyntaxa_table}
\alias{construct_dyntaxa_table}
\title{Construct Dyntaxa taxonomy table from API}
\usage{
construct_dyntaxa_table(parent_ids, subscription_key, shark_output = TRUE)
}
\arguments{
\item{parent_ids}{A list containing parent taxon IDs for which taxonomy information is requested.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{Artdatabanken}.}

\item{shark_output}{Logical. If TRUE, the function will return selected column headers that match SHARK output. If FALSE, all columns are returned. Default is TRUE.}
}
\value{
A data frame representing the constructed taxonomy table.
}
\description{
This function constructs a taxonomy table based on Dyntaxa parent taxon IDs.
It queries the SLU Artdatabanken API (Dyntaxa) to fetch taxonomy information and organizes the data into a hierarchical table.
}
\details{
Please read the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{conditions} and \href{https://api-portal.artdatabanken.se/}{register} before using the API.

Data collected through the API is stored at SLU Artdatabanken.

Genom att anvanda denna applikation atar jag mig att folja regler for
anvandning av SLU Artdatabankens information, inklusive att respektera tredje
mans upphovsratt. Detta atagande gors aven direkt mot SLU. Vid brott mot
detta kan jag forlora ratten att anvanda applikationen
}
\examples{
\dontrun{
# Construct Dyntaxa taxonomy table for taxon IDs 238366 and 1010380
parent_ids <- get_dyntaxa_parent_ids(c(238366, 1010380), "your_subscription_key")
taxonomy_table <- construct_dyntaxa_table(parent_ids, "your_subscription_key")
print(taxonomy_table)
}

}
\seealso{
\href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
