% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{data_deliverer_filter}
\alias{data_deliverer_filter}
\title{Filter datasets based on data deliverer}
\usage{
data_deliverer_filter(datasets, data_deliverer)
}
\arguments{
\item{datasets}{The input datasets to filter.}

\item{data_deliverer}{The data deliverer to filter datasets.}
}
\value{
A filtered data frame.
}
\description{
This internal function filters datasets based on the specified data deliverer.
}
\keyword{internal}
