% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{download_file}
\alias{download_file}
\title{Helper function to download file}
\usage{
download_file(dataset_name)
}
\arguments{
\item{dataset_name}{The name of the dataset to download.}
}
\value{
A temporary file path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This internal function downloads a file based on the specified dataset name.
}
\keyword{internal}
