% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{download_sharkdata}
\alias{download_sharkdata}
\title{Download and process SHARKdata datasets}
\usage{
download_sharkdata(dataset_names)
}
\arguments{
\item{dataset_names}{A character vector specifying the names of the datasets to be downloaded and processed.}
}
\value{
A processed data frame containing the SHARKdata datasets.
}
\description{
This function downloads and processes datasets from the \href{https://sharkdata.smhi.se/}{SHARKdata API} based on the specified dataset names.
}
\details{
This function performs the following steps:
\enumerate{
\item Downloads datasets with the specified names.
\item Reads and processes the downloaded data.
\item Converts columns to appropriate types.
\item Adds Dyntaxa higher taxonomy information to biological data from stored file. Current taxonomy can be accessed via the Dyntaxa API by, see \code{\link{update_dyntaxa_taxonomy}}
\item Adds WoRMS higher taxonomy information to biological data from stored file. Current taxonomy can be accessed via the WoRMS API by, see \code{\link{update_worms_taxonomy}}
\item Extracts and adds geographical information (e.g. location_sea_basin) from coordinates.
}
}
\examples{
\dontrun{
datasets <- load_dataset_names("Phytoplankton")
dataset_names <- unique(datasets$dataset_name)
data <- download_sharkdata(dataset_names)

print(data)
}

}
\seealso{
\code{\link{load_dataset_names}}, \code{\link{update_dyntaxa_taxonomy}}, \code{\link{update_worms_taxonomy}}, \href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
