% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algaebase_api_functions.R
\name{extract_algaebase_field}
\alias{extract_algaebase_field}
\title{Extract specific field from AlgaeBase query result}
\usage{
extract_algaebase_field(query_result, field_name)
}
\arguments{
\item{query_result}{A list object containing the output from an AlgaeBase query.}

\item{field_name}{A character string specifying the field name to extract from the query result.}
}
\value{
A character vector containing the values of the specified field, or \code{NA} if the field is not found.
}
\description{
This helper function extracts a specified field from the list object returned by an AlgaeBase query.
If the requested field is not present, it returns \code{NA} to indicate missing data.
}
\examples{
# Example output from an AlgaeBase query
query_result <- list(species = "Skeletonema marinoi", kingdom = "Chromista")
field_value <- SHARK4R:::extract_algaebase_field(query_result, "species")
print(field_value)

}
\keyword{internal}
