% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{fill_na_below_first_non_na}
\alias{fill_na_below_first_non_na}
\title{Function to fill NA values below the first non-NA value in a vector}
\usage{
fill_na_below_first_non_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector with NAs filled below the first non-NA value.
}
\description{
This internal function fills NA values below the first non-NA value in a vector.
}
\keyword{internal}
