% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{find_descendants}
\alias{find_descendants}
\title{Find all descendants of a taxon}
\usage{
find_descendants(taxon_id, data)
}
\arguments{
\item{taxon_id}{A character string representing the taxon ID for which to find descendants.}

\item{data}{A data frame containing taxonomic information with columns
\code{taxonId} and \code{parentNameUsageID}.}
}
\value{
A tibble containing all descendants of the given taxon, including their
\code{taxonId}, \code{parentNameUsageID}, and any other columns present in the input \code{data}.
}
\description{
This helper function recursively finds all descendant taxa of a given taxon ID
by traversing the parent-child hierarchy in the provided dataset.
}
\examples{
# Example data
data <- dplyr::tibble(
  taxonId = c("1", "2", "3", "4"),
  parentNameUsageID = c(NA, "1", "2", "2"),
  scientificName = c("Root", "Child1", "Child2", "Grandchild1")
)

# Find descendants of taxon "1"
SHARK4R:::find_descendants("1", data)

}
\keyword{internal}
