% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{gather_species_info}
\alias{gather_species_info}
\title{Gather species information}
\usage{
gather_species_info(shark_species_list)
}
\arguments{
\item{shark_species_list}{A data frame containing shark species information.}
}
\value{
A data frame containing gathered species information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This internal function gathers species information from the shark species list.
}
\keyword{internal}
