% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{gather_worms_species_info}
\alias{gather_worms_species_info}
\title{Gather WoRMS species information}
\usage{
gather_worms_species_info(taxa_worms)
}
\arguments{
\item{taxa_worms}{A data frame containing WoRMS species information.}
}
\value{
A data frame containing gathered WoRMS species information.
}
\description{
This internal function gathers WoRMS species information from the taxa_worms data frame.
}
\keyword{internal}
