% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_all_parents}
\alias{get_all_parents}
\title{Get All Parent Taxa for a Set of Taxon IDs}
\usage{
get_all_parents(data, initial_taxon_ids)
}
\arguments{
\item{data}{A data frame containing taxonomic information with columns
\code{taxonId} and \code{parentNameUsageID}. The data can be collected using the
\code{get_dyntaxa_dwca()} function.}

\item{initial_taxon_ids}{A character vector of initial taxon IDs for which
to retrieve all parent taxa. These IDs should correspond to the \code{taxonId}
column in the dataset.}
}
\value{
A tibble containing all parent taxa for the given initial taxon IDs,
including their \code{taxonId}, \code{parentNameUsageID}, and any other columns present
in the input \code{data}.
}
\description{
This helper function iteratively retrieves all parent taxa for a given set
of initial taxon IDs, traversing up the taxonomic hierarchy until the root
(taxa with no parents) is reached.
}
\examples{
# Example data
data <- tibble::tibble(
  taxonId = c("1", "2", "3", "4"),
  parentNameUsageID = c(NA, "1", "1", "2"),
  scientificName = c("Root", "Child1", "Child2", "Grandchild1")
)

# Find all parents of taxon IDs "3" and "4"
SHARK4R:::get_all_parents(data, initial_taxon_ids = c("3", "4"))

}
\keyword{internal}
