% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_dyntaxa_children_hierarchy}
\alias{get_dyntaxa_children_hierarchy}
\title{Get children hierarchies for specified taxon IDs from SLU Artdatabanken API (Dyntaxa)}
\usage{
get_dyntaxa_children_hierarchy(
  taxon_ids,
  subscription_key,
  levels = 1,
  main_children = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{taxon_ids}{A vector of numeric taxon IDs for which children taxon IDs are requested.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{Artdatabanken}.}

\item{levels}{Integer. Default is 1}

\item{main_children}{Logical. Default is TRUE.}

\item{verbose}{Logical. Default is TRUE.}
}
\value{
A data frame containing children taxon information corresponding to the specified taxon IDs.
}
\description{
This function queries the SLU Artdatabanken API (Dyntaxa) to retrieve children taxon hierarchy information for the specified taxon IDs.
It constructs a request with the provided taxon IDs, sends the request to the SLU Artdatabanken API, and
processes the response to return a data frame of taxon children.
}
\details{
\strong{Note}: Please review the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{API conditions}
and \href{https://api-portal.artdatabanken.se/}{register for access} before using the API. Data collected through the API
is stored at SLU Artdatabanken. Please also note that the authors of \code{SHARK4R} are not affiliated with SLU Artdatabanken.
}
\examples{
\dontrun{
# Get children taxon hierarchy for taxon IDs 1010608 and 5000062
children_hierarchy <- get_dyntaxa_children_hierarchy(c(1010608, 5000062), "your_subscription_key")
print(children_hierarchy)
}


}
\seealso{
\href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
\keyword{internal}
