% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_dyntaxa_children_ids}
\alias{get_dyntaxa_children_ids}
\title{Get children taxon IDs for specified taxon IDs from SLU Artdatabanken API (Dyntaxa)}
\usage{
get_dyntaxa_children_ids(
  taxon_ids,
  subscription_key = Sys.getenv("DYNTAXA_KEY"),
  main_children = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{taxon_ids}{A vector of numeric taxon IDs for which children taxon IDs are requested.}

\item{subscription_key}{A Dyntaxa API subscription key. By default, the key
is read from the environment variable \code{DYNTAXA_KEY}.

You can provide the key in three ways:
\itemize{
\item \strong{Directly as a parameter}:
\code{get_dyntaxa_children_ids(1010608, subscription_key = "your_key_here")}
\item \strong{Temporarily for the session}:
\code{Sys.setenv(DYNTAXA_KEY = "your_key_here")}
\item \strong{Permanently across sessions} by adding it to your \code{~/.Renviron} file.
Use \code{usethis::edit_r_environ()} to open the file, then add:
\code{DYNTAXA_KEY=your_key_here}
}}

\item{main_children}{Logical. Default is TRUE.}

\item{verbose}{Logical. Default is TRUE.}
}
\value{
A list containing children taxon IDs corresponding to the specified taxon IDs.
}
\description{
This function queries the SLU Artdatabanken API (Dyntaxa) to retrieve children taxon IDs for the specified taxon IDs.
It constructs a request with the provided taxon IDs, sends the request to the SLU Artdatabanken API, and
processes the response to return a list of children taxon IDs.
}
\details{
A valid Dyntaxa API subscription key is required.
You can request a free key for the "Taxonomy" service from the ArtDatabanken API portal:
\url{https://api-portal.artdatabanken.se/}

\strong{Note}: Please review the \href{https://www.slu.se/artdatabanken/rapportering-och-fynd/oppna-data-och-apier/}{API conditions}
and \href{https://api-portal.artdatabanken.se/}{register for access} before using the API. Data collected through the API
is stored at SLU Artdatabanken. Please also note that the authors of \code{SHARK4R} are not affiliated with SLU Artdatabanken.
}
\examples{
\dontrun{
# Get children taxon IDs for taxon IDs 1010608 and 5000062
children_ids <- get_dyntaxa_children_ids(c(1010608, 5000062), "your_subscription_key")
print(children_ids)
}

}
\seealso{
\href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
\keyword{internal}
