% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_dyntaxa_dwca}
\alias{get_dyntaxa_dwca}
\title{Download and Read Darwin Core Archive Files from Dyntaxa}
\usage{
get_dyntaxa_dwca(subscription_key, file_to_read = "Taxon.csv", verbose = TRUE)
}
\arguments{
\item{subscription_key}{A string containing your API subscription key for the Dyntaxa Taxon Service.}

\item{file_to_read}{A string specifying the name of the CSV file to read from the extracted archive.
Allowed options are: \code{"Reference.csv"}, \code{"SpeciesDistribution.csv"}, \code{"Taxon.csv"}, or \code{"VernacularName.csv"}. Defaults to \code{"Taxon.csv"}.}

\item{verbose}{A logical value indicating whether to show download progress. Defaults to \code{TRUE}.}
}
\value{
A tibble containing the data from the specified CSV file.
}
\description{
This function downloads a complete Darwin Core Archive (DwCA) of Dyntaxa from the SLU Artdatabanken API, extracts the archive, and reads the specified CSV file into R.
}
\details{
\strong{Note}: Please review the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{API conditions}
and \href{https://api-portal.artdatabanken.se/}{register for access} before using the API. Data collected through the API
is stored at SLU Artdatabanken. Please also note that the authors of \code{SHARK4R} are not affiliated with SLU Artdatabanken.
}
\examples{
\dontrun{
# Provide your Dyntaxa API subscription key
subscription_key <- "your_subscription_key"

# Download and read the Taxon.csv file
taxon_data <- get_dyntaxa_dwca(subscription_key, file_to_read = "Taxon.csv")
}

}
