% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_dyntaxa_parent_ids}
\alias{get_dyntaxa_parent_ids}
\title{Get parent taxon IDs for specified taxon IDs from SLU Artdatabanken API (Dyntaxa)}
\usage{
get_dyntaxa_parent_ids(taxon_ids, subscription_key)
}
\arguments{
\item{taxon_ids}{A vector of numeric taxon IDs for which parent taxon IDs are requested.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{Artdatabanken}.}
}
\value{
A list containing parent taxon IDs corresponding to the specified taxon IDs.
}
\description{
This function queries the SLU Artdatabanken API (Dyntaxa) to retrieve parent taxon IDs for the specified taxon IDs.
It constructs a request with the provided taxon IDs, sends the request to the SLU Artdatabanken API, and
processes the response to return a list of parent taxon IDs.
}
\details{
Please read the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{conditions} and \href{https://api-portal.artdatabanken.se/}{register} before using the API.

Data collected through the API is stored at SLU Artdatabanken.

Genom att anvanda denna applikation atar jag mig att folja regler for
anvandning av SLU Artdatabankens information, inklusive att respektera tredje
mans upphovsratt. Detta atagande gors aven direkt mot SLU. Vid brott mot
detta kan jag forlora ratten att anvanda applikationen
}
\examples{
\dontrun{
# Get parent taxon IDs for taxon IDs 238366 and 1010380
parent_ids <- get_dyntaxa_parent_ids(c(238366, 1010380), "your_subscription_key")
print(parent_ids)
}

}
\seealso{
\href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
