% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_dyntaxa_records}
\alias{get_dyntaxa_records}
\title{Get taxonomic information from SLU Artdatabanken API for specified taxon IDs}
\usage{
get_dyntaxa_records(taxon_ids, subscription_key)
}
\arguments{
\item{taxon_ids}{A vector of numeric taxon IDs (Dyntaxa ID) for which taxonomic information is requested.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{Artdatabanken}.}
}
\value{
A data frame containing taxonomic information for the specified taxon IDs.
Columns include taxonId, names, category, rank, isRecommended, and parentTaxonId.
}
\description{
This function queries the SLU Artdatabanken API (Dyntaxa) to retrieve taxonomic information for the specified taxon IDs.
It constructs a request with the provided taxon IDs, sends the request to the SLU Artdatabanken API, and
processes the response to return taxonomic information in a data frame.
}
\details{
Please read the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{conditions} and \href{https://api-portal.artdatabanken.se/}{register} before using the API.

Data collected through the API is stored at SLU Artdatabanken.
}
\examples{
\dontrun{
# Get taxonomic information for taxon IDs 238366 and 1010380
taxon_info <- get_dyntaxa_records(c(238366, 1010380), "your_subscription_key")
print(taxon_info)
}


}
\seealso{
\href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
