% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{get_geographical_info}
\alias{get_geographical_info}
\title{Get geographical information from coordinates}
\usage{
get_geographical_info(latitude_dd, longitude_dd)
}
\arguments{
\item{latitude_dd}{A numeric vector representing latitude coordinates in decimal degrees.}

\item{longitude_dd}{A numeric vector representing longitude coordinates in decimal degrees.}
}
\value{
A data frame containing the original coordinates with the added geographical information location_type_area, location_svar_sea_area_name and location_sea_basin.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This function enriches the provided data with geographical information based on latitude and longitude coordinates in the Baltic Sea, Kattegat and Skagerrak.
}
\details{
The function reads shapefiles and basin names to assign geographical information to the dataset.
It provides details about the location type area and the Svar Sea area name based on the given coordinates.
}
\examples{
\dontrun{
latitude <- c(58.25830, 58.59367, 63.49983)
longitude <- c(11.43330, 18.23550, 19.81933)
dataset_with_geo_info <- get_geographical_info(latitude, longitude)
}

}
\seealso{
\href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
\keyword{internal}
