% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Recursively retrieve parent taxonomic hierarchy}
\usage{
get_hierarchy(taxon_id, data)
}
\arguments{
\item{taxon_id}{A character string representing the taxon ID for which the hierarchy is to be retrieved.}

\item{data}{A data frame containing taxonomic information. Must include columns \code{taxonId}, \code{parentNameUsageID}, and \code{scientificName}.}
}
\value{
A character vector containing the hierarchy of parent \code{scientificName} values, starting with the immediate parent and ascending.
}
\description{
This helper function retrieves the taxonomic hierarchy for a given taxon ID by
recursively traversing its parent relationships in the provided dataset.
}
\examples{
# Example dataset
data <- dplyr::tibble(
  taxonId = c("1", "2", "3"),
  parentNameUsageID = c(NA, "1", "2"),
  scientificName = c("Kingdom", "Phylum", "Class")
)

# Get hierarchy for taxon ID "3"
SHARK4R:::get_hierarchy("3", data)
# [1] "Phylum" "Kingdom"

}
\keyword{internal}
