% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plankton_toolbox.R
\name{get_nomp_list}
\alias{get_nomp_list}
\title{Get the latest NOMP biovolume Excel list}
\usage{
get_nomp_list(
  year = as.numeric(format(Sys.Date(), "\%Y")),
  file = NULL,
  force = FALSE,
  base_url = NULL
)
}
\arguments{
\item{year}{Numeric year to download. Default is current year; if not available,
previous years are automatically tried.}

\item{file}{Character string specifying which file in the zip archive to read.
Defaults to the first Excel file in the archive.}

\item{force}{Logical; if \code{TRUE}, forces re-download of the zip file even if cached copy exists.}

\item{base_url}{Base URL (without "/nomp_taxa_biovolumes_and_carbon_YYYY.zip") for the NOMP biovolume files. Defaults to the SMHI directory.}
}
\value{
A tibble with the contents of the requested Excel file.
}
\description{
This function downloads the latest available NOMP biovolume zip archive
from SMHI (using \code{cache_nomps_zip()}), unzips it, and reads the first
Excel file by default. You can also specify which file in the archive to read.
}
\examples{
\dontrun{
  # Get the latest available list
  nomp_list <- get_nomp_list()

  # Get the 2023 list
  nomp_list_2023 <- get_nomp_list(2023)
}
}
