% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_api_functions.R
\name{get_shark_options}
\alias{get_shark_options}
\title{Retrieve available search options from SHARK API}
\usage{
get_shark_options(prod = TRUE, unparsed = FALSE)
}
\arguments{
\item{prod}{Logical. Query against PROD or TEST (SMHI internal) server. Default is TRUE (PROD).}

\item{unparsed}{Logical. If \code{TRUE}, returns the complete JSON output as list. Defaults to \code{FALSE}.}
}
\value{
A named \code{list} containing the available search options from the SHARK API.
}
\description{
The \code{get_shark_options} function retrieves available search options from the SHARK database.
It sends a GET request to the SHARK API and returns the results as a structured \code{data.frame}.
}
\details{
This function sends a GET request to the SHARK API options endpoint to retrieve available search filters and options
for querying the database.
}
\examples{
\dontrun{
  # Retrieve available search options
  shark_options <- get_shark_options()
  View(shark_options)

  # View available datatypes
  dataTypes <- shark_options$dataTypes
  print(dataTypes)
}

}
\seealso{
\url{https://shark.smhi.se} for SHARK database.

\code{\link{get_shark_data}}
}
