% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_api_functions.R
\name{get_shark_table}
\alias{get_shark_table}
\title{Retrieve and Format SHARK Data from SMHI API}
\usage{
get_shark_table(
  table_view = "sharkweb_overview",
  limit = 200,
  offset = 0,
  header_lang = "internal_key",
  from_year = 2019,
  to_year = 2020,
  months = c(4, 5, 6),
  data_types = c("Chlorophyll", "Epibenthos"),
  prod = TRUE
)
}
\arguments{
\item{table_view}{Character. Specifies the view of the table to retrieve. Options include:
\itemize{
\item \code{"sharkweb_overview"}
\item \code{"sharkweb_all"}
\item \code{"sharkdata_bacterioplankton"}
\item \code{"sharkdata_chlorophyll"}
\item \code{"sharkdata_epibenthos"}
\item \code{"sharkdata_greyseal"}
\item \code{"sharkdata_harbourporpoise"}
\item \code{"sharkdata_harbourseal"}
\item \code{"sharkdata_jellyfish"}
\item \code{"sharkdata_physicalchemical_columns"}
\item \code{"sharkdata_phytoplankton"}
\item \code{"sharkdata_picoplankton"}
\item \code{"sharkdata_planktonbarcoding"}
\item \code{"sharkdata_primaryproduction"}
\item \code{"sharkdata_ringedseal"}
\item \code{"sharkdata_sealpathology"}
\item \code{"sharkdata_sedimentation"}
\item \code{"sharkdata_zoobenthos"}
\item \code{"sharkdata_zooplankton"}
\item \code{"report_sum_year_param"}
\item \code{"report_sum_year_param_taxon"}
\item \code{"report_sampling_per_station"}
\item \code{"report_obs_taxon"}
\item \code{"report_stations"}
\item \code{"report_taxon"}
}
Default is \code{"sharkweb_overview"}.}

\item{limit}{Integer. Maximum number of records to retrieve per request. Default is \code{200}.}

\item{offset}{Integer. The starting point for data retrieval, useful for pagination. Default is \code{0}.}

\item{header_lang}{Character. Language option for column headers. Possible values:
\itemize{
\item \code{"sv"}: Swedish
\item \code{"en"}: English
\item \code{"short"}: Shortened version
\item \code{"internal_key"}: Internal key (default)
}}

\item{from_year}{Integer. The starting year for the data to retrieve. Default is \code{2019}.}

\item{to_year}{Integer. The ending year for the data to retrieve. Default is \code{2020}.}

\item{months}{Integer vector. The months to retrieve data for, e.g., \code{c(4, 5, 6)} for April to June.}

\item{data_types}{Character vector. Specifies data types to filter, such as \code{"Chlorophyll"} or \code{"Epibenthos"}. See \link{get_shark_options} for available data types.}

\item{prod}{Logical. Query against PROD or TEST (SMHI internal) server. Default is TRUE (PROD).}
}
\value{
A \code{data.frame} containing the retrieved data, with column names based on the API's \code{headers}.
Columns are filled with \code{NA} when rows have differing lengths.
}
\description{
The \code{get_shark_table} function retrieves data from the SHARK database hosted by SMHI, with options
to filter based on year range, months, and data types. It sends a POST request to
the SHARK API and returns the results as a structured \code{data.frame}.
}
\details{
This function constructs a JSON body with specified parameters and sends a POST request
to the SHARK API. The API returns data in JSON format, which is then parsed into a \code{data.frame}.
If rows have differing lengths, \code{rbind.fill} fills missing columns with \code{NA}.
}
\examples{
\dontrun{
  # Retrieve chlorophyll data for April to June from 2019 to 2020
  shark_data <- get_shark_table(from_year = 2019, to_year = 2020,
                               months = c(4, 5, 6), data_types = c("Chlorophyll"))
  View(shark_data)
}

}
\seealso{
\code{\link{get_shark_options}}
}
