% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_plot.R
\name{internal_outlier_plot}
\alias{internal_outlier_plot}
\title{Visualize range of data for specific parameters
Uses data from within the dataset to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each dataset}
\usage{
internal_outlier_plot(data, parameter)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
ggplot and tibble of data with outliers
}
\description{
Visualize range of data for specific parameters
Uses data from within the dataset to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each dataset
}
