% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{load_dataset_names}
\alias{load_dataset_names}
\title{Load dataset names from SHARKdata based on type, year, and data deliverer}
\usage{
load_dataset_names(dataset_type, year = NA, data_deliverer = NA)
}
\arguments{
\item{dataset_type}{The type of dataset to load.}

\item{year}{The year to filter datasets based on dataset name (default NA results in all available years). Please note that dataset names can contain multiple years (e.g. a dataset with data from 2012-2014 will not be selected when specifying year 2013 in argument). Consider filter data after download for more extensive data.}

\item{data_deliverer}{The data deliverer (code) to filter datasets (default is NA, results in all deliverers). See \href{http://smhi.se/oceanografi/oce_info_data/shark_web/downloads/codelist_SMHI.xlsx}{SMHI codelist} for available codes. Consider filter data after download for more extensive data.}
}
\value{
A data frame containing filtered dataset names.
}
\description{
This function gets dataset names fron the \href{https://sharkdata.smhi.se/}{SHARKdata API} based on the specified data type, year, and data deliverer.
}
\examples{
\dontrun{
# All Zoobenthos datasets
load_dataset_names("Zoobenthos")

# All Phytoplankton datasets from 2012:2020 from SMHI and UMSC.
load_dataset_names("Phytoplankton", year = 2012:2020, data_deliverer = c("SMHI", "UMSC"))
}

}
\seealso{
\code{\link{load_sharkdata}}, \code{\link{load_dataset_types}}, \href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
