% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{load_dataset_types}
\alias{load_dataset_types}
\title{List available dataset types in SHARKdata}
\usage{
load_dataset_types()
}
\value{
A vector of unique dataset types.
}
\description{
This function lists the available dataset types from the \href{https://sharkdata.smhi.se/}{SHARKdata API}.
}
\examples{
\dontrun{
all_dataset_types <- load_dataset_types()
}

}
\seealso{
\code{\link{load_sharkdata}}, \code{\link{load_dataset_names}}, \href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
