% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stations.R
\name{load_station_bundle}
\alias{load_station_bundle}
\title{Load station database (station.txt) from path, NODC_CONFIG, or package bundle}
\usage{
load_station_bundle(station_file = NULL, verbose = TRUE)
}
\arguments{
\item{station_file}{Optional path to a station.txt file.}

\item{verbose}{Logical; if TRUE, prints messaging about which source is used.}
}
\value{
A data frame containing the station database.
}
\description{
Load station database (station.txt) from path, NODC_CONFIG, or package bundle
}
\keyword{internal}
