% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{load_worms_taxonomy}
\alias{load_worms_taxonomy}
\title{Load higher taxonomy from WoRMS file}
\usage{
load_worms_taxonomy(aphia_id_input)
}
\arguments{
\item{aphia_id_input}{A vector of Aphia IDs to filter the higher taxonomy.}
}
\value{
A data frame containing higher taxonomy information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This internal function loads higher taxonomy information from the stored WoRMS taxonomy file.
}
\seealso{
\code{\link{load_dyntaxa_taxonomy}}, \code{\link{update_dyntaxa_taxonomy}}, \code{\link{update_worms_taxonomy}}, \href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
\keyword{internal}
