% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{match_taxon_name}
\alias{match_taxon_name}
\title{Match Dyntaxa taxon names via API}
\usage{
match_taxon_name(
  taxon_names,
  subscription_key,
  searchFields = "Both",
  isRecommended = "NotSet",
  isOkForObservationSystems = "NotSet",
  culture = "sv_SE",
  page = 1,
  pageSize = 100
)
}
\arguments{
\item{taxon_names}{A vector of taxon names to match.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{Artdatabanken}.}

\item{searchFields}{A character string indicating the search fields. Defaults to 'Both'.}

\item{isRecommended}{A character string indicating whether the taxon is recommended. Defaults to 'NotSet'.}

\item{isOkForObservationSystems}{A character string indicating whether the taxon is suitable for observation systems. Defaults to 'NotSet'.}

\item{culture}{A character string indicating the culture. Defaults to 'sv_SE'.}

\item{page}{An integer specifying the page number for pagination. Defaults to 1.}

\item{pageSize}{An integer specifying the page size for pagination. Defaults to 100.}
}
\value{
A data frame containing the search pattern, taxon ID, and best match for each taxon name.
}
\description{
This function matches a list of taxon names against the SLU Artdatabanken API (Dyntaxa) and retrieves the best matches along with their taxon IDs.
}
\details{
Please read the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{conditions} and \href{https://api-portal.artdatabanken.se/}{register} before using the API.

Data collected through the API is stored at SLU Artdatabanken.

Genom att anvanda denna applikation atar jag mig att folja regler for
anvandning av SLU Artdatabankens information, inklusive att respektera tredje
mans upphovsratt. Detta atagande gors aven direkt mot SLU. Vid brott mot
detta kan jag forlora ratten att anvanda applikationen
}
\examples{
\dontrun{
# Match taxon names against SLU Artdatabanken API
matched_taxa <- match_taxon_name(c("Homo sapiens", "Canis lupus"), "your_subscription_key")
print(matched_taxa)
}

}
\seealso{
\href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
