% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_taxa.R
\name{match_wormstaxa}
\alias{match_wormstaxa}
\title{Taxon matching using WoRMS (http://www.marinespecies.org/)
matches latin name in data with WoRMS taxon list}
\usage{
match_wormstaxa(names, ask = TRUE)
}
\arguments{
\item{names}{Vector of scientific names.}

\item{ask}{Ask user in case of multiple matches.}
}
\value{
Data frame with scientific name, scientific name ID and match type.
}
\description{
Taxon matching using WoRMS (http://www.marinespecies.org/)
matches latin name in data with WoRMS taxon list
}
