% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stations.R
\name{nominal_station}
\alias{nominal_station}
\title{Check if stations are reported as nominal position or not
Function makes an estimated guess whether stations are nominal}
\usage{
nominal_station(data)
}
\arguments{
\item{data}{Data frame.}
}
\value{
Data frame with station name and latitude and longitude positions.
}
\description{
Check if stations are reported as nominal position or not
Function makes an estimated guess whether stations are nominal
}
