% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_plot.R
\name{outlier_plot}
\alias{outlier_plot}
\title{Visualize range of data for specific parameters
Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype}
\usage{
outlier_plot(data, parameter, mild_outlier, extreme_outlier)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
ggplot and tibble of data with outliers
}
\description{
Visualize range of data for specific parameters
Uses data from national marine monitoring for the last 5 years to identify outliers
Ranges and IQR (interquantile range) for specific parameters is adapted to each datatype
}
