% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_leaflet.R
\name{plot_map_leaflet}
\alias{plot_map_leaflet}
\title{Create a Leaflet map.}
\usage{
plot_map_leaflet(data, provider = "Esri.OceanBasemap")
}
\arguments{
\item{data}{The data frame.}

\item{provider}{Tile provider, see
https://leaflet-extras.github.io/leaflet-providers/preview/. Default is \code{Esri.OceanBasemap}.}
}
\value{
HTML widget object.
}
\description{
Create a Leaflet map.
}
