% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{read_data}
\alias{read_data}
\title{Helper function to read data}
\usage{
read_data(temp, filtered_datasets, dataset_name)
}
\arguments{
\item{temp}{The temporary file path.}

\item{filtered_datasets}{The filtered datasets.}

\item{dataset_name}{The name of the dataset.}
}
\value{
A data frame containing the read and converted data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This internal function reads and converts data from a temporary file.
}
\keyword{internal}
