% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{read_species_list}
\alias{read_species_list}
\title{Read and convert species list}
\usage{
read_species_list(filename)
}
\arguments{
\item{filename}{The name of the file containing the species list.}
}
\value{
A data frame containing the species list with converted columns.
}
\description{
This internal function reads and converts a species list from the specified file.
}
\keyword{internal}
