% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read}
\alias{shark_read}
\title{Read tab delimited files downloaded from SHARK
Uses read_delim to read tab delimited files with standardized export format from SHARK}
\usage{
shark_read(filename, delimiters = "point-tab", encoding = "latin_1")
}
\arguments{
\item{filename}{Path to file to be read}

\item{delimiters}{Character. Specifies the delimiter used to separate values in \code{filename}.
Options are \code{"point-tab"} (tab-separated) or \code{"point-semi"} (semicolon-separated).
Default is \code{"point-tab"}.}

\item{encoding}{Character. Specifies the text encoding of \code{filename}.
Options are \code{"cp1252"}, \code{"utf_8"}, \code{"utf_16"}, or \code{"latin_1"}.
Default is \code{"utf_8"}.}
}
\value{
Data frame of file
}
\description{
Read tab delimited files downloaded from SHARK
Uses read_delim to read tab delimited files with standardized export format from SHARK
}
