% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read_deliv}
\alias{shark_read_deliv}
\title{Read files delivered to SHARK
Uses readxl to read excel files with standardized delivery format}
\usage{
shark_read_deliv(filename)
}
\arguments{
\item{filename}{path to file to be read}
}
\value{
Data frame of file
}
\description{
Read files delivered to SHARK
Uses readxl to read excel files with standardized delivery format
}
