% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read_deliv_xls}
\alias{shark_read_deliv_xls}
\title{Read .xls files delivered to SHARK
Uses readxl to read excel files with standardized delivery format}
\usage{
shark_read_deliv_xls(filename, skip = 2, sheet = 2)
}
\arguments{
\item{filename}{path to file to be read}
}
\value{
Data frame of file
}
\description{
Read .xls files delivered to SHARK
Uses readxl to read excel files with standardized delivery format
}
