% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read_zip}
\alias{shark_read_zip}
\title{Read zip archive and unzip tab delimited files downloaded from SHARK}
\usage{
shark_read_zip(zipname)
}
\arguments{
\item{zipname}{path to file to be read}
}
\value{
Data frame of file
}
\description{
Uses unzip (unz) and read_delim to unzip archive and read tab delimited files with standardized export format from SHARK
}
