% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{update_data}
\alias{update_data}
\title{Function to update data if a more recent version is available at SHARKdata}
\usage{
update_data(data)
}
\arguments{
\item{data}{A data frame containing the current data.}
}
\value{
A data frame with updated data if available.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This function updates the data if an updated version is available at \href{https://sharkdata.smhi.se/}{SHARKdata} by downloading the latest datasets.
}
\examples{
# Assume you have a sharkdata dataset named 'data' loaded in your R environment
# with a 'dataset_file_name' column indicating the file name of the dataset.
\dontrun{
datasets <- load_dataset_names("Phytoplankton")
dataset_names <- unique(datasets$dataset_name[1:5])
data <- download_sharkdata(dataset_names)
updated_data <- update_data(data)
}

}
\seealso{
\code{\link{load_dataset_names}}, \code{\link{download_sharkdata}}, \code{\link{check_data_version}}, \href{https://sharkdata.smhi.se/}{SHARKdata}, \href{https://sharkweb.smhi.se/}{SHARKweb}
}
\keyword{internal}
