% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{update_dyntaxa_taxonomy}
\alias{update_dyntaxa_taxonomy}
\title{Update taxonomy from SHARK via SLU Artdatabanken API (Dyntaxa)}
\usage{
update_dyntaxa_taxonomy(
  dyntaxa_ids,
  subscription_key,
  add_missing_taxa = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dyntaxa_ids}{A vector of Dyntaxa taxon IDs to update.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API for Dyntaxa. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken}.}

\item{add_missing_taxa}{Logical. If TRUE, the function will attempt to fetch missing taxa (i.e., taxon_ids not found in the initial Dyntaxa DwC-A query). Default is FALSE.}

\item{verbose}{Logical. Print progress messages. Default is TRUE.}
}
\value{
A data frame representing the updated Dyntaxa taxonomy table.
}
\description{
This function updates Dyntaxa taxonomy records based on a list of Dyntaxa taxon IDs.
It collects parent IDs from SLU Artdatabanken API (Dyntaxa), retrieves full taxonomy records, and organizes
the data into a full taxonomic table that can be joined with data downloaded from \href{https://shark.smhi.se/}{SHARK}
}
\details{
\strong{Note}: Please review the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{API conditions}
and \href{https://api-portal.artdatabanken.se/}{register for access} before using the API. Data collected through the API
is stored at SLU Artdatabanken. Please also note that the authors of \code{SHARK4R} are not affiliated with SLU Artdatabanken.
}
\examples{
\dontrun{
# Update Dyntaxa taxonomy for taxon IDs 238366 and 1010380
updated_taxonomy <- update_dyntaxa_taxonomy(c(238366, 1010380), "your_subscription_key")
print(updated_taxonomy)
}


}
\seealso{
\code{\link{get_shark_data}}, \code{\link{update_worms_taxonomy}}, \href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
