% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms_api_functions.R
\name{update_worms_taxonomy}
\alias{update_worms_taxonomy}
\title{Update taxonomy from SHARKdata datasets via WoRMS API}
\usage{
update_worms_taxonomy(aphiaid)
}
\arguments{
\item{aphiaid}{A numeric vector containing Aphia IDs for which WoRMS taxonomy needs to be updated.}
}
\value{
A data frame containing updated WoRMS taxonomy information.
}
\description{
This function collects WoRMS (World Register of Marine Species) taxonomy information for a given set of Aphia IDs.
The data is organized into a full taxonomic table that can be joined with data downloaded from \href{https://sharkdata.smhi.se/}{SHARKdata}
}
\examples{
\dontrun{
# Update WoRMS taxonomy for a set of Aphia IDs
updated_taxonomy <- update_worms_taxonomy(c(149619, 149122, 11))
print(updated_taxonomy)
}

}
\seealso{
\code{\link{download_sharkdata}}, \code{\link{update_dyntaxa_taxonomy}}, \href{http://www.marinespecies.org/rest/}{WoRMS API Documentation}
}
