% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{validate_dataset_names}
\alias{validate_dataset_names}
\title{Helper function to validate dataset names}
\usage{
validate_dataset_names(filtered_datasets, dataset_names)
}
\arguments{
\item{filtered_datasets}{A data frame of filtered datasets.}

\item{dataset_names}{A vector of dataset names to validate.}
}
\value{
NULL. It stops the execution with an error message if validation fails.
}
\description{
This internal function validates whether all specified dataset names are in the database.
}
\keyword{internal}
