% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{year_filter}
\alias{year_filter}
\title{Filter datasets based on year}
\usage{
year_filter(datasets, year)
}
\arguments{
\item{datasets}{The input datasets to filter.}

\item{year}{The year to filter datasets.}
}
\value{
A filtered data frame.
}
\description{
This internal function filters datasets based on the specified year.
}
\keyword{internal}
