% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_depth.R
\name{check_depth}
\alias{check_depth}
\title{Check which points have potentially invalid depths.}
\usage{
check_depth(data)
}
\arguments{
\item{data}{The data frame.}
}
\value{
Problematic records or an errors report.
}
\description{
Check which points have potentially invalid depths.
}
\details{
The function has been modified from the \code{obistools} package (Provoost and Bosch, 2024).
}
\references{
Provoost P, Bosch S (2024). “obistools: Tools for data enhancement and quality control” Ocean Biodiversity Information System. Intergovernmental Oceanographic Commission of UNESCO. R package version 0.1.0, \url{https://iobis.github.io/obistools/}.
}
