% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_depth.R
\name{check_depth}
\alias{check_depth}
\title{Check which points have potentially invalid depths.}
\usage{
check_depth(data)
}
\arguments{
\item{data}{The data frame.}
}
\value{
Problematic records or an errors report.
}
\description{
Check which points have potentially invalid depths.
}
