% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stations.R
\name{check_station_distance}
\alias{check_station_distance}
\title{Station distance check using SMHI station list "stations.txt" (the list is synced with "Stationsregistret": https://stationsregister.miljodatasamverkan.se/stationsregister/composer/)
matches reported station name in data with curated station list and checks if it is within preset distance}
\usage{
check_station_distance(names, lat, lon)
}
\arguments{
\item{names}{Vector of station names.}

\item{latitude}{and longitude}
}
\value{
Data frame with station name and logical value within/outside preset distance limits.
}
\description{
Station distance check using SMHI station list "stations.txt" (the list is synced with "Stationsregistret": https://stationsregister.miljodatasamverkan.se/stationsregister/composer/)
matches reported station name in data with curated station list and checks if it is within preset distance
}
