% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_logical.R
\name{check_zero_positions}
\alias{check_zero_positions}
\title{Check logical assumptions of data for station positions}
\usage{
check_zero_positions(data)
}
\arguments{
\item{data}{for tibble be be checked}
}
\value{
logical test with true or false results
}
\description{
Check logical assumptions of data for station positions
}
