% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{check_data_version}
\alias{update_data}
\alias{download_sharkdata}
\alias{load_dataset_names}
\alias{load_sharkdata}
\alias{load_dataset_types}
\alias{download_file}
\alias{year_filter}
\alias{data_deliverer_filter}
\alias{read_data}
\alias{validate_dataset_names}
\alias{load_dyntaxa_taxonomy}
\alias{read_species_list}
\alias{gather_species_info}
\alias{add_species_info}
\alias{load_worms_taxonomy}
\alias{gather_worms_species_info}
\alias{add_worms_species_info}
\alias{get_geographical_info}
\alias{filter_outdated_datasets}
\title{Defunct functions}
\usage{
# Deprecated in 0.1.4 -------------------------------------

check_data_version(...)

update_data(...)

download_sharkdata(...)

load_dataset_names(...)

load_sharkdata(...)

load_dataset_types(...)

download_file(...)

year_filter(datasets, year)

data_deliverer_filter(datasets, data_deliverer)

read_data(temp, filtered_datasets, dataset_name)

validate_dataset_names(filtered_datasets, dataset_names)

load_dyntaxa_taxonomy(dyntaxa_id_input)

read_species_list(filename)

gather_species_info(shark_species_list)

add_species_info(shark_species_list, species)

load_worms_taxonomy(aphia_id_input)

gather_worms_species_info(taxa_worms)

add_worms_species_info(taxa_worms, species)

get_geographical_info(latitude_dd, longitude_dd)

filter_outdated_datasets(data, datasets_to_update)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions were deprecated before being made defunct.
If there's a known replacement, calling the function
will tell you about it.
}
\keyword{internal}
