% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\title{Defunct functions}
\usage{
# Deprecated in 1.0.0 -------------------------------------

# Deprecated in 0.1.4 -------------------------------------
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

These functions were deprecated before being made defunct.
If there's a known replacement, calling the function
will tell you about it.
}
\keyword{internal}
