% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{download_file}
\alias{download_file}
\title{Helper function to download file}
\usage{
download_file(dataset_name)
}
\arguments{
\item{dataset_name}{The name of the dataset to download.}
}
\value{
A temporary file path.
}
\description{
This internal function downloads a file based on the specified dataset name.
}
\keyword{internal}
