% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{filter_outdated_datasets}
\alias{filter_outdated_datasets}
\title{Function to filter outdated datasets}
\usage{
filter_outdated_datasets(data, datasets_to_update)
}
\arguments{
\item{data}{A data frame containing the current data.}

\item{datasets_to_update}{A data frame containing information about datasets that need to be updated.}
}
\value{
A filtered data frame excluding outdated datasets.
}
\description{
This internal function filters out outdated datasets from the current data based on the provided list of datasets to update.
}
\keyword{internal}
