% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{gather_species_info}
\alias{gather_species_info}
\title{Gather species information}
\usage{
gather_species_info(shark_species_list)
}
\arguments{
\item{shark_species_list}{A data frame containing shark species information.}
}
\value{
A data frame containing gathered species information.
}
\description{
This internal function gathers species information from the shark species list.
}
\keyword{internal}
