% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharkdata_api_functions.R
\name{gather_worms_species_info}
\alias{gather_worms_species_info}
\title{Gather WoRMS species information}
\usage{
gather_worms_species_info(taxa_worms)
}
\arguments{
\item{taxa_worms}{A data frame containing WoRMS species information.}
}
\value{
A data frame containing gathered WoRMS species information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because SHARKdata has now been replaced by the \href{https://shark.smhi.se/}{SHARK API}. Use \code{\link[=get_shark_table]{get_shark_table()}} instead.

This internal function gathers WoRMS species information from the taxa_worms data frame.
}
\keyword{internal}
