% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_api_functions.R
\name{get_shark_options}
\alias{get_shark_options}
\title{Retrieve Available Search Options from SHARK API}
\usage{
get_shark_options()
}
\value{
A \code{data.frame} containing the available search options from the SHARK API.
}
\description{
The \code{get_shark_options} function retrieves available search options from the SHARK database hosted by SMHI.
It sends a GET request to the SHARK API and returns the results as a structured \code{data.frame}.
}
\details{
This function sends a GET request to the SHARK API options endpoint to retrieve available search filters and options
for querying the database. The API returns data in JSON format, which is then parsed into a \code{data.frame}.
}
\examples{
\dontrun{
  # Retrieve available search options
  shark_options <- get_shark_options()
  View(shark_options)
  
  # View available datatypes
  data_types <- unlist(shark_options$dataTypes)
  print(data_types)
}

}
\seealso{
\code{\link{get_shark_table}}
}
