% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_taxa.R
\name{match_dyntaxa}
\alias{match_dyntaxa}
\title{Taxon matching using Dyntaxa (https://www.dyntaxa.se/)}
\usage{
match_dyntaxa(names)
}
\arguments{
\item{names}{Vector of scientific names.}
}
\value{
Data frame with scientific name, scientific name ID and match type.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because it can now be replaced by using the
\href{https://api-portal.artdatabanken.se/}{Dyntaxa API}. Use \code{\link[=match_taxon_name]{match_taxon_name()}} instead.

matches latin name in data with Dyntaxa taxon list
}
\keyword{internal}
