% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_dyntaxa_taxa.R
\name{match_dyntaxa}
\alias{match_dyntaxa}
\title{Taxon matching using Dyntaxa (https://www.dyntaxa.se/)
matches latin name in data with Dyntaxa taxon list}
\usage{
match_dyntaxa(names)
}
\arguments{
\item{names}{Vector of scientific names.}

\item{ask}{Ask user in case of multiple matches.}
}
\value{
Data frame with scientific name, scientific name ID and match type.
}
\description{
Taxon matching using Dyntaxa (https://www.dyntaxa.se/)
matches latin name in data with Dyntaxa taxon list
}
