% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stations.R
\name{match_station}
\alias{match_station}
\title{Station matching using SMHI station list "station.txt" (the list is synced with "Stationsregistret": https://stationsregister.miljodatasamverkan.se/)
matches reported station name in data with curated station list}
\usage{
match_station(names)
}
\arguments{
\item{names}{Vector of station names.}
}
\value{
Data frame with station name and match type.
}
\description{
Station matching using SMHI station list "station.txt" (the list is synced with "Stationsregistret": https://stationsregister.miljodatasamverkan.se/)
matches reported station name in data with curated station list
}
