% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_leaflet.R
\name{plot_map_leaflet_deliv}
\alias{plot_map_leaflet_deliv}
\title{Create a Leaflet map.}
\usage{
plot_map_leaflet_deliv(data, provider = "Esri.OceanBasemap")
}
\arguments{
\item{data}{The data frame.}

\item{provider}{Tile provider, see
https://leaflet-extras.github.io/leaflet-providers/preview/. Default is \code{Esri.OceanBasemap}.}
}
\value{
HTML widget object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated and has been replaced by \code{\link[=plot_map_leaflet]{plot_map_leaflet()}}.
}
\keyword{internal}
