% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot of user-defined variables and parameters with x- and y-axes automatic selection
Interactive plotly plot}
\usage{
scatterplot(data, ...)
}
\arguments{
\item{data}{for tibble be be checked}

\item{...}{Additional parameters}
}
\value{
plotly plot
}
\description{
Scatterplot of user-defined variables and parameters with x- and y-axes automatic selection
Interactive plotly plot
}
