% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARK4R-package.R
\docType{package}
\name{SHARK4R-package}
\alias{SHARK4R}
\alias{SHARK4R-package}
\title{SHARK4R: SHARK4R: Retrieving, Analyzing, and Validating Marine Data from SHARK and Nordic Microalgae}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package for data analysis, automatic quality control and data retroeval for physical, chemical and marine biological data generated by SHARK and Nordic Microalgae, developed by SMHI. Parts have been modified from https://iobis.github.io/obistools. Provoost P and Bosch S (2018). "obistools: Tools for data enhancement and quality control." Ocean Biogeographic Information System. Intergovernmental Oceanographic Commission of UNESCO. https://cran.r-project.org/package=obistools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sharksmhi/SHARK4R}
  \item \url{https://sharksmhi.github.io/SHARK4R/}
  \item Report bugs at \url{https://github.com/sharksmhi/SHARK4R/issues}
}

}
\author{
\strong{Maintainer}: Anders Torstensson \email{anders.torstensson@smhi.se} (\href{https://orcid.org/0000-0002-8283-656X}{ORCID}) (Swedish Meteorological and Hydrological Institute)

Authors:
\itemize{
  \item Markus Lindh (\href{https://orcid.org/0000-0002-7120-4145}{ORCID}) (Swedish Meteorological and Hydrological Institute)
}

Other contributors:
\itemize{
  \item SHARK \email{shark@smhi.se} [copyright holder]
}

}
\keyword{internal}
