% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARK4R-package.R
\docType{package}
\name{SHARK4R-package}
\alias{SHARK4R}
\alias{SHARK4R-package}
\title{SHARK4R: Data analysis and automatic quality control of physical, chemical and marine biological data generated by the SHARK or Planktontoolbox system}
\description{
R package for data analysis and automatic quality control of physical, chemical and marine biological data generated by the SHARK or Planktontoolbox system developed by Sh at SMHI. Modified from https://iobis.github.io/obistools. Provoost P and Bosch S (2018). "obistools: Tools for data enhancement and quality control." Ocean Biogeographic Information System. Intergovernmental Oceanographic Commission of UNESCO. https://cran.r-project.org/package=obistools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sharksmhi/SHARK4R}
  \item Report bugs at \url{https://github.com/sharksmhi/SHARK4R/issues}
}

}
\author{
\strong{Maintainer}: Markus Lindh \email{first.last@example.com} (\href{https://orcid.org/0000-0002-7120-4145}{ORCID})

Authors:
\itemize{
  \item Anders Torstensson \email{first.last@example.com} (\href{https://orcid.org/0000-0002-8283-656X}{ORCID}) (Swedish Meteorological and Hydrological Institute)
}

Other contributors:
\itemize{
  \item SHARK \email{shark@smhi.se} [copyright holder]
}

}
\keyword{internal}
