% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHARK4R-package.R
\docType{package}
\name{SHARK4R-package}
\alias{SHARK4R}
\alias{SHARK4R-package}
\title{SHARK4R: Accessing and Validating Marine Environmental Data from 'SHARK' and Related Databases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to retrieve, process, analyze, and quality-control marine physical, chemical, and biological data. The main focus is on Swedish monitoring data available through the 'SHARK' database \url{https://shark.smhi.se/en/}, with additional API support for 'Nordic Microalgae' \url{https://nordicmicroalgae.org/}, 'Dyntaxa' \url{https://artfakta.se/}, World Register of Marine Species ('WoRMS') \url{https://www.marinespecies.org}, 'AlgaeBase' \url{https://www.algaebase.org}, OBIS 'xylookup' web service \url{https://iobis.github.io/xylookup/} and Intergovernmental Oceanographic Commission (IOC) - UNESCO databases on harmful algae \url{https://www.marinespecies.org/hab/} and toxins \url{https://toxins.hais.ioc-unesco.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sharksmhi.github.io/SHARK4R/}
  \item \url{https://github.com/sharksmhi/SHARK4R}
  \item Report bugs at \url{https://github.com/sharksmhi/SHARK4R/issues}
}

}
\author{
\strong{Maintainer}: Anders Torstensson \email{anders.torstensson@smhi.se} (\href{https://orcid.org/0000-0002-8283-656X}{ORCID}) (Swedish Meteorological and Hydrological Institute)

Authors:
\itemize{
  \item Markus Lindh (\href{https://orcid.org/0000-0002-7120-4145}{ORCID}) (Swedish Meteorological and Hydrological Institute)
}

Other contributors:
\itemize{
  \item Mikael Hedblom (\href{https://orcid.org/0009-0007-5124-9956}{ORCID}) (Swedish Meteorological and Hydrological Institute) [contributor]
  \item Bengt Karlson (\href{https://orcid.org/0000-0002-7524-3504}{ORCID}) (Swedish Meteorological and Hydrological Institute) [contributor]
  \item SHARK \email{shark@smhi.se} [copyright holder]
  \item SBDI (Swedish Research Council, 2019-00242) [funder]
}

}
\keyword{internal}
