% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read}
\alias{shark_read}
\title{Read tab delimited files downloaded from SHARK
Uses read_delim to read tab delimited files with standardized export format from SHARK}
\usage{
shark_read(filename, type = "txt")
}
\arguments{
\item{filename}{path to file to be read}
}
\value{
Data frame of file
}
\description{
Read tab delimited files downloaded from SHARK
Uses read_delim to read tab delimited files with standardized export format from SHARK
}
