% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read_deliv}
\alias{shark_read_deliv}
\title{Read .xlsx files delivered to SHARK}
\usage{
shark_read_deliv(filename, skip = 2, sheet = 2)
}
\arguments{
\item{filename}{path to file to be read}

\item{skip}{Minimum number of rows to skip before reading anything, be it column names or data. Leading empty rows are automatically skipped, so this is a lower bound. Ignored if range is given. Default is 2.}

\item{sheet}{Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to the second sheet.}
}
\value{
Data frame of file
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated and has been replaced by \code{\link[=read_shark_deliv]{read_shark_deliv()}}.

Uses readxl to read excel files with standardized delivery format
}
\keyword{internal}
