% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shark_read.R
\name{shark_read_zip}
\alias{shark_read_zip}
\title{Read zip archive and unzip tab delimited files downloaded from SHARK
Uses unzip (unz) and read_delim to unzip archive and read tab delimited files with standardized export format from SHARK}
\usage{
shark_read_zip(zipname, type = "zip")
}
\arguments{
\item{zipname}{path to file to be read}

\item{type}{File extension. Default is \code{.zip}.}
}
\value{
Data frame of file
}
\description{
Read zip archive and unzip tab delimited files downloaded from SHARK
Uses unzip (unz) and read_delim to unzip archive and read tab delimited files with standardized export format from SHARK
}
