% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaxa_api_functions.R
\name{update_dyntaxa_taxonomy}
\alias{update_dyntaxa_taxonomy}
\title{Update taxonomy from SHARKdata datasets via SLU Artdatabanken API (Dyntaxa)}
\usage{
update_dyntaxa_taxonomy(dyntaxa_ids, subscription_key)
}
\arguments{
\item{dyntaxa_ids}{A vector of Dyntaxa taxon IDs to update.}

\item{subscription_key}{A character string containing the subscription key for accessing the SLU Artdatabanken API for Dyntaxa. A key is provided for registered users at \href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken}.}
}
\value{
A data frame representing the updated Dyntaxa taxonomy table.
}
\description{
This function updates Dyntaxa taxonomy records based on a list of Dyntaxa taxon IDs.
It collects parent IDs from SLU Artdatabanken API (Dyntaxa), retrieves full taxonomy records, and organizes
the data into a full taxonomic table that can be joined with data downloaded from \href{https://sharkdata.smhi.se/}{SHARKdata}
}
\details{
Please read the \href{https://www.artdatabanken.se/tjanster-och-miljodata/oppna-data-och-apier/api-villkor/}{conditions} and \href{https://api-portal.artdatabanken.se/}{register} before using the API.

Data collected through the API is stored at SLU Artdatabanken.

Genom att anvanda denna applikation atar jag mig att folja regler for
anvandning av SLU Artdatabankens information, inklusive att respektera tredje
mans upphovsratt. Detta atagande gors aven direkt mot SLU. Vid brott mot
detta kan jag forlora ratten att anvanda applikationen
}
\examples{
\dontrun{
# Update Dyntaxa taxonomy for taxon IDs 238366 and 1010380
updated_taxonomy <- update_dyntaxa_taxonomy(c(238366, 1010380), "your_subscription_key")
print(updated_taxonomy)
}

}
\seealso{
\code{\link{download_sharkdata}}, \code{\link{update_worms_taxonomy}}, \href{https://api-portal.artdatabanken.se/}{SLU Artdatabanken API Documentation}
}
